Renamed project to reflect on the Wiki bot name. Updated composer and files with the new namespace
