+++ composer.json
@@ -5,7 +5,7 @@
 
     "autoload" : {
         "psr-4" : {
-            "GiphyBot\\" : "app/GiphyBot"
+            "WikiToSimple\\" : "app/WikiToSimple"
         }
     }
 }
