+++ index.php
@@ -1,11 +1,11 @@
 <?php
 
-use GiphyBot\Controller\Reddit;
-use GiphyBot\Config;
+use WikiToSimple\Controller\Reddit;
+use WikiToSimple\Config;
 
 require_once("vendor/autoload.php");
 
-require_once("app/GiphyBot/Config/Settings.php");
+require_once("app/WikiToSimple/Config/Settings.php");
 
 $reddit = new Reddit($settings['username'], $settings['password'], $settings['clientId'], $settings['clientSecret']);
 $isConnected = $reddit->connect();
